create job seeker basic apis

+++ src/main/java/com/example/consultantapp/ServiceImpl/UserServiceImpl.java
@@ -1,10 +1,13 @@
 package com.example.consultantapp.ServiceImpl;
 
+import com.example.consultantapp.DTO.JobSeekerDTO;
 import com.example.consultantapp.DTO.LoginDetailsDTO;
 import com.example.consultantapp.DTO.UserDTO;
 import com.example.consultantapp.DTO.UserFullDTO;
+import com.example.consultantapp.Model.JobSeeker;
 import com.example.consultantapp.Model.LoginDetails;
 import com.example.consultantapp.Model.User;
+import com.example.consultantapp.Repository.LoginDetailsRepository;
 import com.example.consultantapp.Repository.UserRepository;
 import com.example.consultantapp.Service.UserService;
 import org.modelmapper.ModelMapper;
@@ -19,21 +22,34 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private UserRepository userRepository;
     @Autowired
+    private LoginDetailsRepository loginDetailsRepository;
+    @Autowired
     private LoginDetailsServiceImpl loginDetailsServiceImpl;
+    @Autowired
+    private JobSeekerServiceImpl jobSeekerServiceImpl;
     @Override
     public UserDTO addUser(UserFullDTO userData) {
         try{
-         User user = modelMapper.map(userData, User.class);
-         User newUser = userRepository.save(user);
-         if(newUser!=null){
-             LoginDetailsDTO ld = loginDetailsServiceImpl.addLoginDetails(userData,user);
-             if(ld!=null){
-                 return modelMapper.map(user, new TypeToken<UserDTO>() {}.getType());
-             }
-             else{
-                 userRepository.deleteByUserIdEquals(newUser.getUserId());
-             }
-         }
+            boolean validEmail = loginDetailsServiceImpl.validateEmail(userData.getEmail());
+            if(validEmail){
+                 User user = modelMapper.map(userData, User.class);
+                 User newUser = userRepository.save(user);
+                 if(newUser!=null){
+                     LoginDetailsDTO ld = loginDetailsServiceImpl.addLoginDetails(userData,user);
+                     if(ld!=null){
+                         JobSeekerDTO jobSeeker = jobSeekerServiceImpl.registerJobSeeker(newUser);
+                         if(jobSeeker!=null){
+                             return modelMapper.map(user, new TypeToken<UserDTO>() {}.getType());
+                         }
+                         else{
+                             userRepository.deleteByUserIdEquals(newUser.getUserId());
+                         }
+                     }
+                     else{
+                         userRepository.deleteByUserIdEquals(newUser.getUserId());
+                     }
+                 }
+            }
          return null;
         }
         catch(Exception e){
